+++ staging/src/k8s.io/apiserver/pkg/util/flowcontrol/fairqueuing/event_clock.go
@@ -20,7 +20,6 @@ import (
 	"container/heap"
 	"math/rand"
 	"sync"
-	"testing"
 	"time"
 
 	"k8s.io/apimachinery/pkg/util/clock"
@@ -33,17 +32,20 @@ import (
 // no other work is left to be completed in goroutines.
 type EventFunc func(time.Time)
 
+// SettablePassiveClock allows setting current time of a passive clock
 type SettablePassiveClock interface {
 	clock.PassiveClock
 	SetTime(time.Time)
 }
 
+// EventClock fires event on time
 type EventClock interface {
 	clock.PassiveClock
 	EventAfterDuration(f EventFunc, d time.Duration)
 	EventAfterTime(f EventFunc, t time.Time)
 }
 
+// RealEventClock fires event on real world time
 type RealEventClock struct {
 	clock.RealClock
 }
@@ -76,7 +78,8 @@ type FakeEventClock struct {
 	clock.FakePassiveClock
 
 	// waiters is a heap of waiting work, sorted by time
-	waiters eventWaiterHeap
+	waiters     eventWaiterHeap
+	waitersLock sync.RWMutex
 
 	// clientWG may be nil and if not supplies constraints on time
 	// passing in Run.  The Run method will not pick a new time until
@@ -125,6 +128,8 @@ func NewFakeEventClock(t time.Time, clientWG *sync.WaitGroup, fuzz time.Duration
 // GetNextTime returns the next time at which there is work scheduled,
 // and a bool indicating whether there is any such time
 func (fec *FakeEventClock) GetNextTime() (time.Time, bool) {
+	fec.waitersLock.RLock()
+	defer fec.waitersLock.RUnlock()
 	if len(fec.waiters) > 0 {
 		return fec.waiters[0].targetTime, true
 	}
@@ -154,28 +159,34 @@ func (fec *FakeEventClock) Run(limit *time.Time) {
 func (fec *FakeEventClock) SetTime(t time.Time) {
 	fec.FakePassiveClock.SetTime(t)
 	for {
-		// This loop is because events run at a given time may schedule more
-		// events to run at that or an earlier time.
-		// Events should not advance the clock.  But just in case they do...
-		now := fec.Now()
-		var wg sync.WaitGroup
 		foundSome := false
-		for len(fec.waiters) > 0 && !now.Before(fec.waiters[0].targetTime) {
-			ew := heap.Pop(&fec.waiters).(eventWaiter)
-			wg.Add(1)
-			go func(f EventFunc) { f(now); wg.Done() }(ew.f)
-			foundSome = true
-		}
+		func() {
+			fec.waitersLock.Lock()
+			defer fec.waitersLock.Unlock()
+			// This loop is because events run at a given time may schedule more
+			// events to run at that or an earlier time.
+			// Events should not advance the clock.  But just in case they do...
+			now := fec.Now()
+			var wg sync.WaitGroup
+			for len(fec.waiters) > 0 && !now.Before(fec.waiters[0].targetTime) {
+				ew := heap.Pop(&fec.waiters).(eventWaiter)
+				wg.Add(1)
+				go func(f EventFunc) { f(now); wg.Done() }(ew.f)
+				foundSome = true
+			}
+			wg.Wait()
+		}()
 		if !foundSome {
 			break
 		}
-		wg.Wait()
 	}
 }
 
 // EventAfterDuration schedules the given function to be invoked once
 // the given duration has passed.
 func (fec *FakeEventClock) EventAfterDuration(f EventFunc, d time.Duration) {
+	fec.waitersLock.Lock()
+	defer fec.waitersLock.Unlock()
 	now := fec.Now()
 	fd := time.Duration(float32(fec.fuzz) * fec.rand.Float32())
 	heap.Push(&fec.waiters, eventWaiter{targetTime: now.Add(d + fd), f: f})
@@ -184,6 +195,8 @@ func (fec *FakeEventClock) EventAfterDuration(f EventFunc, d time.Duration) {
 // EventAfterTime schedules the given function to be invoked once
 // the given time has arrived.
 func (fec *FakeEventClock) EventAfterTime(f EventFunc, t time.Time) {
+	fec.waitersLock.Lock()
+	defer fec.waitersLock.Unlock()
 	fd := time.Duration(float32(fec.fuzz) * fec.rand.Float32())
 	heap.Push(&fec.waiters, eventWaiter{targetTime: t.Add(fd), f: f})
 }
@@ -204,27 +217,4 @@ func (ewh *eventWaiterHeap) Pop() interface{} {
 	x := old[n-1]
 	*ewh = old[:n-1]
 	return x
-}
-
-func exercisePassiveClock(t *testing.T, pc SettablePassiveClock) {
-	t1 := time.Now()
-	t2 := t1.Add(time.Hour)
-	pc.SetTime(t1)
-	tx := pc.Now()
-	if tx != t1 {
-		t.Errorf("SetTime(%#+v); Now() => %#+v", t1, tx)
-	}
-	dx := pc.Since(t1)
-	if dx != 0 {
-		t.Errorf("Since() => %v", dx)
-	}
-	pc.SetTime(t2)
-	dx = pc.Since(t1)
-	if dx != time.Hour {
-		t.Errorf("Since() => %v", dx)
-	}
-	tx = pc.Now()
-	if tx != t2 {
-		t.Errorf("Now() => %#+v", tx)
-	}
 }
